Allow applying additional patches
Not very pretty, but useful when having to apply local patches for the
Python code. Can be done like this:

    (self: super: {
      inherit (authentik-nix.lib.mkAuthentikScope {
        pkgs = self;
        extraPatches = [
          ./hackhack.patch
        ];
      }) authentikComponents;
    })

This cannot be used currently for Go patches since I left it out for the
gopkgs part intentionally due to the filterSource stuff since it's
somewhat likely that a patch contains stuff that's filtered out by the
`filterSource` expression. Not sure what to do about that.